+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - A new method, `sendWithDetails`, is exposed on the `XrpClient` and `XpringClient`,
 allowing memo data to be attached to the transaction when sending XRP.
 
+#### Removed
+- `Utils.byteArrayToHex` is deprecated.  Please use `CommonUtils.byteArrayToHex`.
+- `Utils.hexStringToByteArray` is deprecated.  Please use `CommonUtils.stringToByteArray`.
+
 ## 6.0.2 - 2020-08-01
 
 This release contains updated dependencies for stability and security.